requirements.txt
anyio==4.8.0
Authlib==1.4.0
autopep8==2.3.1
blinker==1.9.0
Bottleneck
certifi==2024.12.14
cffi==1.17.1
click
cloudpickle==3.1.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dill==0.3.9
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
fsspec==2024.12.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multiprocess==0.70.17
networkx==3.4.2
numexpr
numpy==1.26.4
packaging==24.2
pandas
pillow==11.1.0
prompt_toolkit==3.0.48
psutil==6.1.1
pycodestyle==2.12.1
pycparser==2.22
pyparsing==3.2.1
python-dateutil
python-dotenv
pytz
schwab-py==1.4.0
scikit-learn
scipy
Shimmy==2.0.0
six
sniffio==1.3.1
stable_baselines3==2.4.1
sympy==1.13.1
threadpoolctl
tomli==2.2.1
torch==2.5.1
typing_extensions==4.12.2
tzdata
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
